requirements_gpu.txt
json_lines
transformers
tensorflow-gpu
